// src/pages/ETicket/types.ts
export type ETicketProp = {
  bookCode: string;
  departureDate: string;
  arrivalDate: string;
  flightCode: string;
  seatClass: string;
  airlineName: string;
  departAirportName: string;
  departIata: string;
  departCity: string;
  departProvince: string;
  arrivalAirportName: string;
  arrivalIata: string;
  arrivalCity: string;
  arrivalProvince: string;
  departureTime: string;
  arrivalTime: string;
};